+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Changed Docker images repository from DockerHub to Amazon ECR.
 - Fixed off-by-one error in virtio-block descriptor address validation.
+- Removed the `flush` capability from the block device
 
 ## [0.24.0]